test(api-core): move switch find tests to integration suite
