+++ crates/api-core/src/tests/switch_find.rs
@@ -21,163 +21,6 @@ use crate::tests::common::api_fixtures::create_test_env;
 use crate::tests::common::api_fixtures::site_explorer::new_switch;
 
 #[crate::sqlx_test]
-async fn test_find_switch_ids_and_by_ids(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let switch_id1 = new_switch(&env, Some("Switch1".to_string()), None).await?;
-    let switch_id2 = new_switch(&env, Some("Switch2".to_string()), None).await?;
-
-    // FindSwitchIds should return both switches
-    let switch_ids = env
-        .api
-        .find_switch_ids(tonic::Request::new(rpc::forge::SwitchSearchFilter {
-            ..Default::default()
-        }))
-        .await?
-        .into_inner()
-        .ids;
-    assert!(switch_ids.contains(&switch_id1));
-    assert!(switch_ids.contains(&switch_id2));
-
-    // FindSwitchesByIds should return the requested switch
-    let switches = env
-        .api
-        .find_switches_by_ids(tonic::Request::new(rpc::forge::SwitchesByIdsRequest {
-            switch_ids: vec![switch_id1],
-        }))
-        .await?
-        .into_inner()
-        .switches;
-    assert_eq!(switches.len(), 1);
-    assert_eq!(switches[0].id, Some(switch_id1));
-
-    // FindSwitchesByIds should return both when requested
-    let switches = env
-        .api
-        .find_switches_by_ids(tonic::Request::new(rpc::forge::SwitchesByIdsRequest {
-            switch_ids: vec![switch_id1, switch_id2],
-        }))
-        .await?
-        .into_inner()
-        .switches;
-    assert_eq!(switches.len(), 2);
-
-    Ok(())
-}
-
-// The empty-list and over-max guards for `find_switches_by_ids` are shared
-// API-layer code, proven once across representative RPCs in
-// `tests::find_by_ids_guards`.
-
-#[crate::sqlx_test]
-async fn test_find_switch_ids_excludes_deleted(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let switch_id1 = new_switch(&env, Some("Switch1".to_string()), None).await?;
-    let switch_id2 = new_switch(&env, Some("Switch2".to_string()), None).await?;
-
-    // Delete switch2
-    env.api
-        .delete_switch(tonic::Request::new(rpc::forge::SwitchDeletionRequest {
-            id: Some(switch_id2),
-        }))
-        .await?;
-
-    // FindSwitchIds should only return the non-deleted switch
-    let switch_ids = env
-        .api
-        .find_switch_ids(tonic::Request::new(rpc::forge::SwitchSearchFilter {
-            ..Default::default()
-        }))
-        .await?
-        .into_inner()
-        .ids;
-    assert!(switch_ids.contains(&switch_id1));
-    assert!(!switch_ids.contains(&switch_id2));
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_find_switch_ids_deleted_only(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let switch_id1 = new_switch(&env, Some("Switch1".to_string()), None).await?;
-    let switch_id2 = new_switch(&env, Some("Switch2".to_string()), None).await?;
-
-    env.api
-        .delete_switch(tonic::Request::new(rpc::forge::SwitchDeletionRequest {
-            id: Some(switch_id2),
-        }))
-        .await?;
-
-    // DELETED_FILTER_ONLY (1) should return only the deleted switch
-    let switch_ids = env
-        .api
-        .find_switch_ids(tonic::Request::new(rpc::forge::SwitchSearchFilter {
-            deleted: 1,
-            ..Default::default()
-        }))
-        .await?
-        .into_inner()
-        .ids;
-    assert!(!switch_ids.contains(&switch_id1));
-    assert!(switch_ids.contains(&switch_id2));
-
-    // DELETED_FILTER_INCLUDE (2) should return both
-    let switch_ids = env
-        .api
-        .find_switch_ids(tonic::Request::new(rpc::forge::SwitchSearchFilter {
-            deleted: 2,
-            ..Default::default()
-        }))
-        .await?
-        .into_inner()
-        .ids;
-    assert!(switch_ids.contains(&switch_id1));
-    assert!(switch_ids.contains(&switch_id2));
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_find_switch_ids_by_controller_state(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let switch_id = new_switch(&env, Some("Switch1".to_string()), None).await?;
-
-    // New switches start in "created" state -- filter for it
-    let switch_ids = env
-        .api
-        .find_switch_ids(tonic::Request::new(rpc::forge::SwitchSearchFilter {
-            controller_state: Some("created".to_string()),
-            ..Default::default()
-        }))
-        .await?
-        .into_inner()
-        .ids;
-    assert!(switch_ids.contains(&switch_id));
-
-    // Filter for a state that doesn't match
-    let switch_ids = env
-        .api
-        .find_switch_ids(tonic::Request::new(rpc::forge::SwitchSearchFilter {
-            controller_state: Some("ready".to_string()),
-            ..Default::default()
-        }))
-        .await?
-        .into_inner()
-        .ids;
-    assert!(!switch_ids.contains(&switch_id));
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
 async fn test_find_switches_by_ids_response_fields(
     pool: sqlx::PgPool,
 ) -> Result<(), Box<dyn std::error::Error>> {
