+++ crates/api-core/tests/integration/main.rs
@@ -28,4 +28,5 @@ mod power_shelf_find;
 mod rack_find;
 mod route_servers;
 mod static_address_management;
+mod switch_find;
 mod tenant_keyset_find;
